Refactoring completed

+++ src/main/java/com/github/savely03/context/Calculator.java
@@ -2,6 +2,8 @@ package com.github.savely03.context;
 
 import com.github.savely03.strategy.CalculatorStrategy;
 
+import java.math.BigDecimal;
+
 public class Calculator {
     private CalculatorStrategy calculatorStrategy;
 
@@ -9,7 +11,7 @@ public class Calculator {
         this.calculatorStrategy = calculatorStrategy;
     }
 
-    public double calculate(int num1, int num2) {
+    public BigDecimal calculate(double num1, double num2) {
         return calculatorStrategy.calculate(num1, num2);
     }
 }
